Fix 64-bit pointer truncation in C interface wrappers
All occam-to-C wrapper functions that receive workspace parameters
used 'int *w' or 'int *ws'. On 64-bit targets workspace slots are
8 bytes but int is 4 bytes, so reading pointer-valued arguments via
w[N] truncates 64-bit pointers to 32 bits.

Changed to 'intptr_t *w' in 12 files across bsclib (proclib,
cspdrvlib, filelib), raster, sdlraster, pony, cdx, udc, video,
and the runtime bsyscalls.

+++ modules/bsclib/libsrc/cspdrvlib/cspdrvc.c
@@ -17,6 +17,7 @@
  *	Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -434,11 +435,11 @@ static __inline__ void cspdrv_release (int *addr)
 
 
 /* occam entry points */
-void _cspdrvlib_init (int *ws)
+void _cspdrvlib_init (intptr_t *ws)
 {
 	cspdrv_init ((char *)(ws[0]), (int)(ws[1]), (int)(ws[2]), (int *)(ws[3]));
 }
-void _cspdrvlib_release (int *ws)
+void _cspdrvlib_release (intptr_t *ws)
 {
 	cspdrv_release ((int *)(ws[0]));
 }
